Implemented basic functionality of CLI and Hierarchy.
Now populates the Hierarchy with a root Ent and an Animals Ent, then
listens for commands via the console.

Tested commands are "f" which displays current focus, and "c" which
lists children. Changing focus is bugged as of now.

Most everything so far should be documented sufficiently well with
comments.