New class/file which handles the analysis of Ents in a Hierarchy.
First implemented is an algorithm to analyze the children of a given Ent
and determine if they reference each other. Each child must either overlap,
or be exclusive to another. If this is not the case for two Ents with
the same direct parent, then one of those Ents must be a parent of the
other, as per the inherent logic of the taxonomical hierarchy.

Class method returns pairs of these "estranged" siblings which require
the user to take action in order to have a complete record of information.