Implemented system which prompts user when two children of an Ent do not reference each other as either exclusive or overlapping. The current term for this situation calls these two Ents "estranged children".
The user will be asked whether the two children overlap or are
exclusive, and if neither, it will ask which is the parent of the other.
One of these options must be the case as dictated by the logic of the
taxonomical hierarchy.

Many new algorithms were added in order to properly analyze the
aforementioned situations.