Refactored some stuff.
Added a Tree handler class called TreeInstance. Many of its functions,
including the constructor and destructor, are private to prevent direct access by subclasses of the abstract EntsInterface.

TreeInstance is friends with the abstract class EntsInterface. A subclass of EntsInterface can use inherited protected functions to request
pointers to instances of TreeInstance.

Public functions of TreeInstance allow for it to be treated as an
object.