Fair amount of refactoring to clear up forward declaration problems.
For the purposes of encapsulation, wrapper classes were added for Tree
and Ent, called TreeInstance and EntInstance. This allows subclasses of
EntsInterface to use the wrappers as objects representing the abstract
concepts, but without being exposed to the ability to alter their
properties.

To alter the properties of an Ent within a Tree, subclasses must call
protected functions inherited from EntsInterface. These functions will
ensure validity of actions. When the safe functions from EntsInterface
must get more information from the user, it is done so via virtual
functions implemented by the subclass, which perform specific things
like query the user for text with short instructions.

The Tests static class was added to provide testing for the legality of
Tree and Ent names.