Added the ability to rename and print the name of Trees.
Added command "clear" to print out a lot of newlines.

Changed the command of "n" to have no argument. The creation of new Ents
is now handled by protected EntsInterface functions which call on
virtual methods of subclasses in order to dictate protocol.

Help command can no longer be reached by "h", as that is valuable for
other purposes perhaps. Now reachable via "help", "-h" and "--help",
just in case people use either.

Added virtual function in EntsInterface for displaying general messages
to the user. This allows the Ents database to interact with the user
in a more direct way to ensure proper knowledge and usage.

When the subclasses of EntsInterface wish to perform an action to change
a property of the tree or its nodes, it will make a request via
protected functions to ensure that such actions are valid.

Removed some commands that have yet to be re-implemented. They will be
added later, but prehaps in different ways.