Changed a number of functions in the Ent class which return lists to send their info by value. There was no reason to send by pointer, as the lists should be quite small, except perhaps ancestors and descendents, but those can be handled differently later if need be.
Before it was sent by a pointer to the data, and it had to be released
manually. This was a bad idea. It makes it harder for memory management,
and if we give a pointer to the original information, then it can be
altered via dereferencing. We instead want that information to only
be changed via special functions.